check latestRelease is null outside of client, return current task instead of cancel, remove argument matcher
